+++ paddle/phi/kernels/gpu/adamw_kernel.cu
@@ -49,12 +49,12 @@ __global__ void AdamWKernelREG(MT beta1,
                                T* param_out,
                                const MT* master_param,
                                MT* master_param_out,
-                               int ndim) {
+                               int64_t ndim) {
   MT lr = *lr_ * lr_ratio;
   MT beta1_pow = beta1_pow_;
   MT beta2_pow = beta2_pow_;
 
-  int id = blockIdx.x * blockDim.x + threadIdx.x;
+  int64_t id = blockIdx.x * blockDim.x + threadIdx.x;
 
   for (; id < ndim; id += gridDim.x * blockDim.x) {
     MT p = master_param ? master_param[id] : static_cast<MT>(param[id]);
@@ -98,12 +98,12 @@ __global__ void AdamWKernelMEM(MT beta1,
                                T* param_out,
                                const MT* master_param,
                                MT* master_param_out,
-                               int ndim) {
+                               int64_t ndim) {
   MT lr = *lr_ * lr_ratio;
   MT beta1_pow = *beta1_pow_;
   MT beta2_pow = *beta2_pow_;
 
-  int id = blockIdx.x * blockDim.x + threadIdx.x;
+  int64_t id = blockIdx.x * blockDim.x + threadIdx.x;
 
   for (; id < ndim; id += gridDim.x * blockDim.x) {
     MT p = master_param ? master_param[id] : static_cast<MT>(param[id]);
